+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-## 19/01/2018 - Version 0.1.0
+## 22/01/2018 - Version 1.0.0
+
+Release changes:
+
+* Implement GraphQLCursor.
+* Support for GraphQL query's and mutations.
+* `GraphQLClient.query` now returns a GraphQLCursor.
+
+## 19/01/2018 - Version 2.0.0
 
 Release changes: